Major Procedure Explanation textimporter refactor
Procedure expl. text importer extends PopupTextImporter.
Change switch statement to  user selection in main menu.
Implement ingeritance across text importers
